implement get records endpoint

+++ src/main/scala/btc/wallet/models/Transaction.scala
@@ -2,4 +2,4 @@ package btc.wallet.models
 
 import java.sql.Timestamp
 
-final case class Transaction(dateTime: Timestamp, amount: Int)
+final case class Transaction(dateTime: Timestamp, amount: Float)
